Initial attempt at supporting client re-authentication
The server-side might be done, but we won't know until we add the
client-side logic to actually try a re-authentication. This
solution is just a guess at the moment, though I suspect it has a
reasonable chance of at least being close to correct.
Client-side almost done; just need KafkaChannel.reauthenticate()

At this point the channel has been told to re-authenticate and
replace the given ExpiringCredential on or after the given time.
The given time may be now/in the past, in which case the
re-authentication attempt can begin now, or it may be in the
future, in which case it should not start until later.
Regardless, we are now at the point where we can (somehow) inject
the re-authentication requests into the channel and process the
corresponding responses, and then we simply have to notify the
ChannelCredentialMgr instance about whether the attempt succeeds
or fails.  That's the last remaining piece -- how to inject the
requests.
Add AuthenticationRequestSender -- not done yet, but close...
Potentially done -- just need to test.
Re-auth works for KafkaConsumer and ControllerChannelManager

Still have a few others to fix, but we're getting there...
Refactoring, started KafkaAdminClient
Fix syntax errors
Fix syntax error
fix findbugs/checkstyle issues
Reactoring, better Javadoc, eliminate whitespace changes
More Javadoc changes and minor adjustments
Add sasl.login.refresh.reauthenticate.enable
Add org.apache.kafka.common.security.expiring.ExpiringCredential