Respond to review comments.
Eliminated supportsClientReauth() and supportsServerReauth() from
Authenticator by adding clientSessionReauthenticationTimeMs() and
serverSessionExpirationTimeNanos().  These new methods only ever
return non-null on the client and server, respectively.
clientSessionReauthenticationTimeMs() now contains the 85%-95%
calculation that was in KafkaChannel. This eliminated the need to
expose a session begin time on the Authenticator, which in turn
eliminated the need for Time (and any changes) in PLAINTEXT and
SSL authenticators.